[MOD-14988] tag_index: put WritePostingsDelta before the tests, tag_value_tests into tests/
Moves WritePostingsDelta/write_postings up next to the TagIndex struct they
belong to, so the file ends in its test modules (Rust convention). Moves
TagValue's own unit tests to tests/integration/tag_value.rs, since they only
touch the public API and can run as an integration test. Documents, above
the remaining three unit-test modules, why each of *those* stays
crate-internal: each reaches a private item (TrieLookup's tuple field,
SuffixWildcardPattern::anchor, expansion_deadline) an integration test can't.

Also suppresses a should_implement_trait warning on TagValue::from_str with
a #[expect] and reason: it deliberately isn't FromStr (Option, not Result).
